requirements.txt
streamlit
pytesseract
opencv-python
Pillow
numpy
pandas
matplotlib
scikit-learn
tqdm
pytest
paddlepaddle
paddleocr
easyocr